# src/backend/go-crud-api_new/utils/bruno/openapi.yaml
openapi: 3.0.0
info:
  title: Go CRUD API
  description: API documentation for the Go CRUD application.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Local server
paths:
  /users:
    get:
      summary: Get all users
      operationId: getUsers
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        description: User object that needs to be added to the database
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /plant:
    post:
      summary: Create a new plant
      operationId: createPlant
      requestBody:
        description: Plant object that needs to be added to the database
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPlant'
      responses:
        '201':
          description: Plant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    NewUser:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
...
    Plant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          type: string
        userId:
          type: string
    NewPlant:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
...